Rewrote input framework, wrote unit test skeleton for input
