+++ src/support/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources(glge
 	PRIVATE
-		${CMAKE_CURRENT_LIST_DIR}/event/input/input_controller.cpp
+		${CMAKE_CURRENT_LIST_DIR}/event/input/input_dispatcher.cpp
 		${CMAKE_CURRENT_LIST_DIR}/proc_gen/terrain.cpp
         #${CMAKE_CURRENT_LIST_DIR}/proc_gen/tree.cpp
 )
