+++ test/CMakeLists.txt
@@ -30,6 +30,7 @@ add_quick_test(stepper)
 add_quick_test(circular)
 add_quick_test(oscillator)
 add_quick_test(events)
+add_quick_test(input)
 add_quick_test(unique_handle)
 add_quick_test(parse_model)
 add_quick_test(parse_model_big)
